Modified   Controls/LEDBase.h Modified   Controls/Rotary.h Modified   DSPs/Chorus.h Modified   DSPs/Flanger.h Modified   DSPs/Phaser.h Modified   DSPs/Wah.h Modified   Filters/TripleToneControlFilter.h Modified   IHAL.h

+++ IHAL.h
@@ -39,8 +39,8 @@ public:
 
 	virtual void InitializePersistentData(uint16 ID) = 0;
 	virtual bool ContainsPersistentData(uint16 ID) = 0;
-	virtual void SetPersistentData(uint16 ID, const void *const Data, uint8 Size) = 0;
-	virtual void GetPersistentData(uint16 ID, void *Data, uint8 Size) = 0;
+	virtual void SetPersistentData(uint16 ID, const void *const Data, uint16 Size) = 0;
+	virtual void GetPersistentData(uint16 ID, void *Data, uint16 Size) = 0;
 	virtual void EreasPersistentData(void) = 0;
 	virtual void SavePersistentData(void) = 0;
